perf(dsv41): fuse CSA2 metadata preparation
Fuse bounded addresses, decode metadata, selection output and compressor
pairs; reuse candidate rows and native BF16 attention all-reduce.

B300 microbench, 16 rows: selection epilogue (top-k 512, candidates 128)
37.12 -> 2.23 us; compressor metadata 52.36 -> 1.43 us.

+++ python/tokenspeed/runtime/models/deepseek_v41.py
@@ -799,11 +799,11 @@ class DeepseekV41Attention(nn.Module):
         out, _ = self.wo_b(out, scale=None)
         if self.mapping.attn.has_tp:
             out = all_reduce(
-                out.float(),
+                out,
                 group=self.mapping.attn.tp_group,
                 backend=None,
                 op=torch.distributed.ReduceOp.SUM,
-            ).to(hidden_states.dtype)
+            )
         return out
